Fix: Eliminate unnecessary and bogus file copy
The function convert_file() was copying the file into the temporary
folder for the sass compiler to find its include files. Since we have
now configured the sass compiler to consider both the cache folder and
the original folder as potential roots, this is no longer necessary.

The removed code even had a bug where the string "scss" in the path
occurred once too many times, which is also resolved with this commit.